#11-dashboard: Added jsCookies Parsher And Created A Dashboard page
