+++ package.json
@@ -32,6 +32,7 @@
     "clsx": "^1.0.4",
     "express": "^4.17.1",
     "history": "^5.0.0",
+    "js-cookies": "^1.0.4",
     "jss": "^10.0.0-alpha.21",
     "jss-rtl": "^0.2.3",
     "next": "9.0.2",
